+++ packages/spacecat-shared-data-access/src/dto/site.js
@@ -11,6 +11,8 @@
  */
 
 import { createSite } from '../models/site.js';
+import AuditConfig from '../models/site/audit-config.js';
+import Config from '../models/site/config.js';
 
 /**
  * Data transfer object for Site.
@@ -25,11 +27,13 @@ export const SiteDto = {
     id: site.getId(),
     baseURL: site.getBaseURL(),
     gitHubURL: site.getGitHubURL() || '',
-    imsOrgId: site.getImsOrgId() || '',
+    organizationId: site.getOrganizationId() || '',
     isLive: site.isLive(),
     createdAt: site.getCreatedAt(),
     updatedAt: site.getUpdatedAt(),
     GSI1PK: 'ALL_SITES',
+    auditConfig: AuditConfig.toDynamoItem(site.getAuditConfig()),
+    config: Config.toDynamoItem(site.getConfig()),
   }),
 
   /**
@@ -42,10 +46,12 @@ export const SiteDto = {
       id: dynamoItem.id,
       baseURL: dynamoItem.baseURL,
       gitHubURL: dynamoItem.gitHubURL,
-      imsOrgId: dynamoItem.imsOrgId,
+      organizationId: dynamoItem.organizationId,
       isLive: dynamoItem.isLive,
       createdAt: dynamoItem.createdAt,
       updatedAt: dynamoItem.updatedAt,
+      auditConfig: dynamoItem.auditConfig,
+      config: dynamoItem.config,
     };
 
     return createSite(siteData);