+++ packages/spacecat-shared-data-access/CHANGELOG.md
@@ -1,3 +1,80 @@
+# [@adobe/spacecat-shared-data-access-v1.5.2](https://github.com/adobe-rnd/spacecat-shared/compare/@adobe/spacecat-shared-data-access-v1.5.1...@adobe/spacecat-shared-data-access-v1.5.2) (2023-12-30)
+
+
+### Bug Fixes
+
+* **deps:** update external fixes ([#76](https://github.com/adobe-rnd/spacecat-shared/issues/76)) ([5b74623](https://github.com/adobe-rnd/spacecat-shared/commit/5b74623a4accfb601b607170f9bd67fb380f5c84))
+
+# [@adobe/spacecat-shared-data-access-v1.5.1](https://github.com/adobe-rnd/spacecat-shared/compare/@adobe/spacecat-shared-data-access-v1.5.0...@adobe/spacecat-shared-data-access-v1.5.1) (2023-12-27)
+
+
+### Bug Fixes
+
+* include audit ref and audited at ([#75](https://github.com/adobe-rnd/spacecat-shared/issues/75)) ([e573c90](https://github.com/adobe-rnd/spacecat-shared/commit/e573c903851d88c7293caa5e5a9af763987edb3a))
+
+# [@adobe/spacecat-shared-data-access-v1.5.0](https://github.com/adobe-rnd/spacecat-shared/compare/@adobe/spacecat-shared-data-access-v1.4.4...@adobe/spacecat-shared-data-access-v1.5.0) (2023-12-26)
+
+
+### Features
+
+* store previous latest audit result ([#74](https://github.com/adobe-rnd/spacecat-shared/issues/74)) ([4663c43](https://github.com/adobe-rnd/spacecat-shared/commit/4663c4344e260f85c4e64691e0685260d36629e5))
+
+# [@adobe/spacecat-shared-data-access-v1.4.4](https://github.com/adobe-rnd/spacecat-shared/compare/@adobe/spacecat-shared-data-access-v1.4.3...@adobe/spacecat-shared-data-access-v1.4.4) (2023-12-25)
+
+
+### Bug Fixes
+
+* add missing pattern to interface ([#72](https://github.com/adobe-rnd/spacecat-shared/issues/72)) ([0eacbff](https://github.com/adobe-rnd/spacecat-shared/commit/0eacbff39dbc12d604e40ae94285533f7804ca0a))
+
+# [@adobe/spacecat-shared-data-access-v1.4.3](https://github.com/adobe-rnd/spacecat-shared/compare/@adobe/spacecat-shared-data-access-v1.4.2...@adobe/spacecat-shared-data-access-v1.4.3) (2023-12-23)
+
+
+### Bug Fixes
+
+* **deps:** update external fixes ([#71](https://github.com/adobe-rnd/spacecat-shared/issues/71)) ([b9a6679](https://github.com/adobe-rnd/spacecat-shared/commit/b9a667966e7b0f013f46c62f6cc4a85f80623c61))
+
+# [@adobe/spacecat-shared-data-access-v1.4.2](https://github.com/adobe-rnd/spacecat-shared/compare/@adobe/spacecat-shared-data-access-v1.4.1...@adobe/spacecat-shared-data-access-v1.4.2) (2023-12-20)
+
+
+### Bug Fixes
+
+* **deps:** update aws-sdk-js-v3 monorepo to v3.477.0 ([#68](https://github.com/adobe-rnd/spacecat-shared/issues/68)) ([492913f](https://github.com/adobe-rnd/spacecat-shared/commit/492913f0011d557b7c91ba631c13ced171bc2ce9))
+
+# [@adobe/spacecat-shared-data-access-v1.4.1](https://github.com/adobe-rnd/spacecat-shared/compare/@adobe/spacecat-shared-data-access-v1.4.0...@adobe/spacecat-shared-data-access-v1.4.1) (2023-12-18)
+
+
+### Bug Fixes
+
+* update dynamo client ([#65](https://github.com/adobe-rnd/spacecat-shared/issues/65)) ([cc09041](https://github.com/adobe-rnd/spacecat-shared/commit/cc09041200dfe23c0a83c384fc9695ac972564e9))
+
+# [@adobe/spacecat-shared-data-access-v1.4.0](https://github.com/adobe-rnd/spacecat-shared/compare/@adobe/spacecat-shared-data-access-v1.3.2...@adobe/spacecat-shared-data-access-v1.4.0) (2023-12-16)
+
+
+### Features
+
+* update audit config ([#63](https://github.com/adobe-rnd/spacecat-shared/issues/63)) ([00a468e](https://github.com/adobe-rnd/spacecat-shared/commit/00a468e9404676212ec9441c8a7fccf1a93c636f))
+
+# [@adobe/spacecat-shared-data-access-v1.3.2](https://github.com/adobe-rnd/spacecat-shared/compare/@adobe/spacecat-shared-data-access-v1.3.1...@adobe/spacecat-shared-data-access-v1.3.2) (2023-12-16)
+
+
+### Bug Fixes
+
+* GSI sort key when error ([#62](https://github.com/adobe-rnd/spacecat-shared/issues/62)) ([dbb6a91](https://github.com/adobe-rnd/spacecat-shared/commit/dbb6a91f75fa9492a621de7b258cbebe38fe3ac7))
+
+# [@adobe/spacecat-shared-data-access-v1.3.1](https://github.com/adobe-rnd/spacecat-shared/compare/@adobe/spacecat-shared-data-access-v1.3.0...@adobe/spacecat-shared-data-access-v1.3.1) (2023-12-16)
+
+
+### Bug Fixes
+
+* type def / jsdoc ([#61](https://github.com/adobe-rnd/spacecat-shared/issues/61)) ([078f9e6](https://github.com/adobe-rnd/spacecat-shared/commit/078f9e6855d3f475f3a746561758aaefb5cb3c79))
+
+# [@adobe/spacecat-shared-data-access-v1.3.0](https://github.com/adobe-rnd/spacecat-shared/compare/@adobe/spacecat-shared-data-access-v1.2.7...@adobe/spacecat-shared-data-access-v1.3.0) (2023-12-16)
+
+
+### Features
+
+* add audit config ([#59](https://github.com/adobe-rnd/spacecat-shared/issues/59)) ([f851862](https://github.com/adobe-rnd/spacecat-shared/commit/f8518623a8b8e4e337b3c44a16b65724a953596a))
+
 # [@adobe/spacecat-shared-data-access-v1.2.7](https://github.com/adobe-rnd/spacecat-shared/compare/@adobe/spacecat-shared-data-access-v1.2.6...@adobe/spacecat-shared-data-access-v1.2.7) (2023-12-11)